requirements.txt
# Data & ML
duckdb
pandas
numpy
scipy
scikit-learn>=1.3
kmodes
ruptures
pyyaml
tsfresh

# Visualization
matplotlib
seaborn

# Deep learning
torch>=2.0
lightning>=2.0
keras
tensorflow

# Config management
hydra-core>=1.3
omegaconf>=2.3

# Experiment tracking
wandb
